+++ map.h
@@ -16,13 +16,13 @@
 typedef struct s_map	t_map;
 typedef struct s_map
 {
-	int				key;
+	long				key;
 	char			*value;
 	t_map			*next;
 }					t_map;
 
-t_map	*create(int key, char *value);
-t_map	*add(t_map *map, int key, char *value);
-char	*find(t_map *map, int key);
+t_map	*create(long key, char *value);
+t_map	*add(t_map *map, long key, char *value);
+char	*find(t_map *map, long key);
 
 #endif
